backend/requirements.txt
django
djangorestframework
tensorflow
pillow
numpy
django-cors-headers
